+++ docs/components/Divider/Web.stories.tsx
@@ -11,6 +11,7 @@ export default {
   component: Divider,
   parameters: {
     viewMode: "story",
+    previewTabs: { code: { hidden: false } },
   },
 } as ComponentMeta<typeof Divider>;